backend/pyproject.toml
[project]
name = "spaceship-backend"
version = "0.1.0"
description = "AI-orchestrated logistics analytics backend"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "duckdb>=1.1",
    "pandas>=2.2",
    "numpy>=1.26",
    "pydantic>=2.9",
    "openai>=1.55",
    "python-dotenv>=1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "httpx>=0.27",
]

[tool.uv]
package = false
